Fix PR mistakes
